+++ lib/src/workspace/src/sensor_grid.dart
@@ -33,17 +33,81 @@ class SensorGridContent with _$SensorGridContent {
 class SensorGridItem with _$SensorGridItem {
   const factory SensorGridItem({
     /// Represents the Asset ID
-    String? assetId,
+    String? objectId,
+
+    /// [kind] represents the kind of card should be displayed
+    @WorkspaceCardTypeConverter() required WorkspaceCardType kind,
 
     /// [dimensions] Represents the dimension of the sensor
-    required GridDimension dimensions,
+    GridDimension? dimensions,
 
     /// Represents the color of the card
     @ColorConverter() @Default(Colors.white) Color color,
 
     /// Represents the sensors of the card
-    required List<SensorGridContent> sensors,
+    @Default([]) List<SensorGridContent> sensors,
   }) = _SensorGridItem;
 
   factory SensorGridItem.fromJson(Map<String, dynamic> json) => _$SensorGridItemFromJson(json);
+}
+
+enum WorkspaceCardType {
+  /// [asset] represents a classic sensors grid card, used to display sensors
+  /// Layrz API reference: `ASSET`
+  asset,
+
+  /// [inboundService] represents a card used to display inbound services
+  /// Layrz API reference: `INBOUND_SERVICE`
+  inboundService,
+
+  /// [outboundService] represents a card used to display outbound services
+  /// Layrz API reference: `OUTBOUND_SERVICE`
+  outboundService,
+  ;
+
+  @override
+  String toString() => toJson();
+
+  String toJson() {
+    switch (this) {
+      case WorkspaceCardType.asset:
+        return 'ASSET';
+      case WorkspaceCardType.inboundService:
+        return 'INBOUND_SERVICE';
+      case WorkspaceCardType.outboundService:
+        return 'OUTBOUND_SERVICE';
+    }
+  }
+
+  static WorkspaceCardType fromJson(String json) {
+    switch (json) {
+      case 'INBOUND_SERVICE':
+        return WorkspaceCardType.inboundService;
+      case 'OUTBOUND_SERVICE':
+        return WorkspaceCardType.outboundService;
+      case 'ASSET':
+      default:
+        return WorkspaceCardType.asset;
+    }
+  }
+}
+
+class WorkspaceCardTypeConverter implements JsonConverter<WorkspaceCardType, String> {
+  const WorkspaceCardTypeConverter();
+
+  @override
+  WorkspaceCardType fromJson(String json) => WorkspaceCardType.fromJson(json);
+
+  @override
+  String toJson(WorkspaceCardType object) => object.toJson();
+}
+
+class WorkspaceCardTypeOrNullConverter implements JsonConverter<WorkspaceCardType?, String?> {
+  const WorkspaceCardTypeOrNullConverter();
+
+  @override
+  WorkspaceCardType? fromJson(String? json) => json != null ? WorkspaceCardType.fromJson(json) : null;
+
+  @override
+  String? toJson(WorkspaceCardType? object) => object?.toJson();
 }